Fix python path errors on startup in newly added trsrn sim configs
Looks like the build process messes with the python vismach simulation files in the config folders. I'm not sure why that happens but moving the files relevant to the vismach simulation models into a new subfolder inside the config folder seems to fix it.
