+++ configs/sim/axis/vismach/5axis/table-rotary_spindle-rotary-nutating/vismach/xyzbca-trsrn-gui.py
@@ -1,9 +1,13 @@
 #!/usr/bin/env python3
 
+# This is a machine simulation for the 'xyzbca-trsrn' simulation config in linuxcnc 
+# Author: David mueller 
+# email: mueller_david@hotmail.com
+
+from twp_vismach import *
 import hal
 import math
 import sys
-
 import os
 
 # getting the name of the directory
@@ -19,7 +23,6 @@ sys.path.append(parent)
 # directory.
 from twp_vismach import *
 
-
 # Machine zero as measured from center surface of the rotary c table
 machine_zero_x = -1000
 machine_zero_y =  1000
@@ -33,12 +36,12 @@ try: # Expect files in working directory
 
 except Exception as detail:
     print(detail)
-    raise SystemExit("gen-xyzacb-trsrn-gui requires stl files in working directory")
+    raise SystemExit("xyzbca-trsrn-gui requires stl files in working directory")
 '''
 
 for setting in sys.argv[1:]: exec(setting)
 
-c = hal.component("xyzacb-trsrn-gui")
+c = hal.component("xyzbca-trsrn-gui")
 # axis_x
 c.newpin("axis_x", hal.HAL_FLOAT, hal.HAL_IN)
 # axis_y
@@ -57,13 +60,13 @@ c.newpin("nutation_angle", hal.HAL_FLOAT, hal.HAL_IN)
 c.newpin("tool_length", hal.HAL_FLOAT, hal.HAL_IN)
 c.newpin("tool_diameter", hal.HAL_FLOAT, hal.HAL_IN)
 # geometric offsets in the spindle-rotary-assembly
-c.newpin("pivot_y", hal.HAL_FLOAT, hal.HAL_IN)
+c.newpin("pivot_x", hal.HAL_FLOAT, hal.HAL_IN)
 c.newpin("pivot_z", hal.HAL_FLOAT, hal.HAL_IN)
 c.newpin("offset_x", hal.HAL_FLOAT, hal.HAL_IN)
 c.newpin("offset_y", hal.HAL_FLOAT, hal.HAL_IN)
 # rot-point offsets distances from pivot point ( spindle AB)
 # to rotation axis ( table C)
-c.newpin("rot_axis_y", hal.HAL_FLOAT, hal.HAL_IN)
+c.newpin("rot_axis_x", hal.HAL_FLOAT, hal.HAL_IN)
 c.newpin("rot_axis_z", hal.HAL_FLOAT, hal.HAL_IN)
 # selected kinematics
 c.newpin("kinstype_select", hal.HAL_FLOAT, hal.HAL_IN)
@@ -123,15 +126,15 @@ class HalSpindleHousingZ(CylinderZ):
         # start the spindle housing at hight 100 above the spindle nose
         return length, r, 100, r
 
-# used to create the spindle rotary as set by the variable 'pivot_y'
-class HalSpindleHousingY(CylinderY):
+# used to create the spindle rotary as set by the variable 'pivot_x'
+class HalSpindleHousingX(CylinderX):
     def __init__(self, comp, *args):
         CylinderZ.__init__(self, *args)
         self.comp = c
 
     def coords(self):
         r = 120
-        length =  c.pivot_y + 150 
+        length =  c.pivot_x + 150 
         return length, r-30, 0, r
 
 # used to create an indicator for the variable 'pivot_z'
@@ -146,15 +149,15 @@ class HalPivotZ(CylinderZ):
         # start the spindle housing at hight 100 above the spindle nose
         return length, 1, 0, r
 
-# used to create an indicator for the variable 'pivot_y'
-class HalPivotY(CylinderY):
+# used to create an indicator for the variable 'pivot_x'
+class HalPivotX(CylinderX):
     def __init__(self, comp, *args):
         CylinderZ.__init__(self, *args)
         self.comp = c
 
     def coords(self):
         r = 2
-        length = -c.pivot_y
+        length = -c.pivot_x
         return length, r, 0, 1
 
 # used to create an indicator for the variable 'offset_x'
@@ -221,8 +224,8 @@ class HalRotateNutation(HalVectorRotate):
 
     def get_values(self):
         th = self.th
-        x = 0
-        y = th*sin(radians(c.nutation_angle))
+        x = th*sin(radians(c.nutation_angle))
+        y = 0
         z = th*cos(radians(c.nutation_angle))
 
         return th, x, y, z
@@ -245,7 +248,7 @@ work_coords_tcp = HalShow([work_coords],c,1,"kinstype_select")
 # work coordinates for tool mode
 work_coords_tool = HalShow([work_coords],c,2,"kinstype_select")
 # rotate to match the current tool orientation 
-work_coords_tool = HalRotateNutation([work_coords_tool],c,"rotary_b",-1,0,0,0)
+work_coords_tool = HalRotateNutation([work_coords_tool],c,"rotary_a",-1,0,0,0)
 work_coords_tool = HalRotate([work_coords_tool],c,"rotary_c",1,0,0,1)
 
 # move the coords to the designated (x,y,z)-home position
@@ -278,11 +281,11 @@ coords = Collection([
 
 # create a visual for the position of the rotational axis of the rotary C table
 # as set by the user with the "rot-axis-x" and "rot-axis-y" sliders in the axis gui
-rot_axis = Color([1,1,0.3,1],[CylinderX(0,2,-2000,2)])
+rot_axis = Color([1,1,0.3,1],[CylinderY(0,2,2000,2)])
 # move the rot_axis indicator by the designated (x,y)-home offset
-rot_axis= Translate([rot_axis], 0, machine_zero_y, machine_zero_z)
+rot_axis= Translate([rot_axis], machine_zero_x, 0, machine_zero_z)
 # now move it to the position set by the user 
-rot_axis = HalVectorTranslate([rot_axis],c,0,"rot_axis_y","rot_axis_z")
+rot_axis = HalVectorTranslate([rot_axis],c,"rot_axis_x",0,"rot_axis_z")
 #/indicators
 
 #tool-side
@@ -303,10 +306,10 @@ tool = Collection([
 tool = Color([1,0,1,0], [tool] )
 # create an indicator for the variable pivot_z
 ind_pivot_z = Color([0,0,0.75,1],[HalPivotZ(CylinderZ)])
-# create an indicator for the variable pivot_y
-ind_pivot_y = Color([0,0.75,0,1],[HalPivotY(CylinderY)])
-# move the pivot_y indicator so it extends from the pivot point
-ind_pivot_y = HalVectorTranslate([ind_pivot_y],c,0,"pivot_y","pivot_z",1)
+# create an indicator for the variable pivot_x
+ind_pivot_x = Color([0.75,0,0,1],[HalPivotX(CylinderX)])
+# move the pivot_x indicator so it extends from the pivot point
+ind_pivot_x = HalVectorTranslate([ind_pivot_x],c,"pivot_x",0,"pivot_z",1)
 # create the spindle housing that contains the spindle
 spindle_housing = Collection([
                   # spindle
@@ -322,8 +325,8 @@ spindle_housing = Collection([
                   ind_pivot_z
                   ])
 
-# spindle housing horizontal length as set by 'pivot_y'
-spindle_housing_horizontal = Color([0.7,0.7,0,1],[HalSpindleHousingY(c)])
+# spindle housing horizontal length as set by 'pivot_x'
+spindle_housing_horizontal = Color([0.7,0.7,0,1],[HalSpindleHousingX(c)])
 # move up along the spindle housing
 spindle_housing_horizontal = Translate([spindle_housing_horizontal],0,0,50)
 # move farther up by the pivot_z value
@@ -334,11 +337,11 @@ spindle_housing_horizontal = HalShow([spindle_housing_horizontal],c,True,"hide_s
 spindle_housing = Collection([
                   spindle_housing,
                   spindle_housing_horizontal,
-                  ind_pivot_y
+                  ind_pivot_x
                   ])
 # move the spindle and it's vertical housing by the values set for the pivot lengths
 # so the vismach origin is in the pivot point
-spindle_housing = HalVectorTranslate([spindle_housing],c,0,"pivot_y","pivot_z",-1)
+spindle_housing = HalVectorTranslate([spindle_housing],c,"pivot_x",0,"pivot_z",-1)
 
 # create a visual for the rotary c axis
 ind_axis_c = Color([1,1,0.3,1],[CylinderZ(0,1,400,1)])
@@ -352,16 +355,16 @@ ind_pivot_point = Collection([
                   ind_axis_c
                   ])
 
-# create the part of joint b that attaches to the spindle body 
+# create the part of joint A that attaches to the spindle body 
 spindle_housing_nut = Collection([
                       Color([0.7,0.7,0,1],[CylinderZ(0,60,150,145)]),
                       Color([0.7,0.7,0,1],[CylinderZ(150,145,180,145)])
                       ])
 # rotate the nutation joint to the nutation angle
-spindle_housing_nut = HalRotate([spindle_housing_nut],c,"nutation_angle",-1,1,0,0)
+spindle_housing_nut = HalRotate([spindle_housing_nut],c,"nutation_angle",1,0,1,0)
 spindle_housing_nut = HalShow([spindle_housing_nut],c,True,"hide_spindle_body",0,1)
 # create HAL-link for a-axis rotational joint"
-spindle_housing = HalRotateNutation([spindle_housing],c,"rotary_b",-1,0,1,0)
+spindle_housing = HalRotateNutation([spindle_housing],c,"rotary_a",-1,0,0,0)
 # move the joint using the offset values as set by the user in the gui panel
 spindle_housing = HalVectorTranslate([spindle_housing],c,"offset_x","offset_y",0,-1)
 spindle_housing = Collection([
@@ -375,9 +378,9 @@ ind_offset_x =  Color([1,0,0,1],[HalOffsetX(c)])
 ind_offset_y =  Color([0,1,0,1],[HalOffsetY(c)])
 ind_offset_y = HalTranslate([ind_offset_y],c,"offset_x", -1,0,0)
 
-# create an indicator for rotational axis B
+# create an indicator for rotational axis A
 ind_axis_a = Color([1,1,0.3,1],[CylinderZ(0,1,400,1)])
-ind_axis_a = HalRotate([ind_axis_a],c,"nutation_angle",-1,1,0,0)
+ind_axis_a = HalRotate([ind_axis_a],c,"nutation_angle",1,0,1,0)
 # move the indicator by the offset values as set by the user in the gui panel
 ind_axis_a = HalVectorTranslate([ind_axis_a],c,"offset_x","offset_y",0,-1)
 spindle_indicators = Collection([
@@ -387,18 +390,20 @@ spindle_indicators = Collection([
                  ind_axis_a,
                  ])
 
-# create part of joint B that attaches to the body of rotary c
+# create part of joint A that attaches to the body of rotary c
 spindle_rotary_nut = Collection([
                      Color([1,0.5,0,0],[CylinderZ(180,145,300,145)]),
                      Color([1,0.5,0,0],[CylinderZ(300,145,310,125)]),
                      ])
 # rotate the nutation joint to the nutation angle
-spindle_rotary_nut = HalRotate([spindle_rotary_nut],c,"nutation_angle",-1,1,0,0)
+spindle_rotary_nut = HalRotate([spindle_rotary_nut],c,"nutation_angle",1,0,1,0)
 
 # create the part that connects the nutation joint to the rotary C 
 spindle_rotary_arm = Color([1,0.5,0,0],[BoxCenteredXY(300, 600, 100)])
 spindle_rotary_arm = Rotate([spindle_rotary_arm],50,-1,0,0)
 spindle_rotary_arm = Translate([spindle_rotary_arm],0,120,200)
+# rotate so it slants in the machine-x direction
+spindle_rotary_arm = Rotate([spindle_rotary_arm],90,0,0,-1)
 # create where the plate the arm connects to the xyz-slide
 spindle_rotary_plate = Color([1,0.5,0,0],[CylinderZ(0,160,50,180)])
 # move the plate up where it connects to the arm
@@ -431,7 +436,7 @@ spindle_xyz = Collection([
              ])
 # move the spindle and it's vertical housing by the values set for the pivot lengths
 # so the vismach origin is in the center of the spindle nose
-spindle_xyz = HalVectorTranslate([spindle_xyz],c,0,"pivot_y","pivot_z")
+spindle_xyz = HalVectorTranslate([spindle_xyz],c,"pivot_x",0,"pivot_z")
 spindle_xyz = HalVectorTranslate([spindle_xyz],c,"offset_x","offset_y",0)
 # spindle head and y-slide move with x
 spindle_xyz = HalTranslate([spindle_xyz],c,"axis_x",1,0,0)
@@ -453,12 +458,8 @@ work_piece = Color([0.5,0.5,0.5,0.9], [work_piece])
 # move the workpiece 
 work_piece = Translate([work_piece],0,300,300)
 work_piece = HalShow([work_piece],c,1,"hide_work_piece_1",0,1)
-# create a block that holds the rotary a table
-rotary_table_housing = Color([0.2,0.2,0.2,0],[BoxCenteredXY(2000, 2000, -1000)])
-# move so the surface set below from the current XY plane
-rotary_table_housing = Translate([rotary_table_housing],0,0,-20)
-# create rotary_table_a
-rotary_table_a = Color([0.1,0.7,0.9,0],[CylinderZ(0,920,-800,920)])
+# create rotary_table_c
+rotary_table_b = Color([0.1,0.7,0.9,0],[CylinderZ(0,920,-800,920)])
 # create 'slots' in x and y on the rotary table
 table_console = Color([0.5,0.5,0.5,0.9],[BoxCenteredXY(1840, 200, 1200)])
 # move so the surface is on the rotation axis
@@ -467,27 +468,21 @@ slot_z_w = Color([1,1,1,1],[BoxCenteredXY(10, 1840, 0.9)])
 # create indicator of at positive end of x slot
 slot_pocket = Color([1,1,1,1],[CylinderZ(0,20,1,20)])
 slot_pocket = Translate([slot_pocket],900,0,0)
-rotary_table_a = Collection([
+rotary_table_b = Collection([
                  work,
                  work_piece,
-                 rotary_table_a,
+                 rotary_table_b,
                  table_console,
-                 slot_z_w
+                 slot_z_w,
                  ])
-# create HAL-link for a-axis rotational joint"
-rotary_table_a = HalRotate([rotary_table_a],c,"rotary_a",-1,0,0,1)
-rotary_table_a = Collection([
-                 rotary_table_a,
-                 rotary_table_housing
-                 ])
-# rotate into the yz plane
-rotary_table_a = Rotate([rotary_table_a],90,0,1,0)
-# rotate so the table_console is horizontal
-rotary_table_a = Rotate([rotary_table_a],90,1,0,0)
-rotary_table_a = Translate([rotary_table_a ],-1700,0,0)
+# rotate into the xz plane
+rotary_table_b = Rotate([rotary_table_b],90,1,0,0)
+# create HAL-link for rotary table 
+rotary_table_b = HalRotate([rotary_table_b],c,"rotary_b",-1,0,1,0)
+rotary_table_b = Translate([rotary_table_b ],0,1700,0)
 
 # rotate the tcp coords with the rotary table
-work_coords_tcp = HalRotate([work_coords_tcp],c,"rotary_a",-1,1,0,0)
+work_coords_tcp = HalRotate([work_coords_tcp],c,"rotary_b",-1,0,1,0)
 
 # create the work_plane using the origin (as measured from current work offset (world),
 # normal vector vz(zx,zy,zz) and  x-orientation vector vx(xx,xy,xz) 
@@ -519,7 +514,7 @@ work_plane = Translate([work_plane], machine_zero_x, machine_zero_y, machine_zer
 # move plane to current work offset 
 work_plane = HalVectorTranslate([work_plane],c,"twp_ox_world","twp_oy_world","twp_oz_world")
 table = Collection([
-        rotary_table_a,
+        rotary_table_b,
         rot_axis,
         work_coords_tcp,
         coords,
@@ -542,7 +537,7 @@ model = Collection([table, spindle_xyz, base,])
 
 #hud
 myhud = Hud()
-myhud.show("XYZACB-trsrn")
+myhud.show("XYZBCA-trsrn")
 myhud.show("----------------")
 myhud.show("TWP-Status:")
 myhud.add_txt("Undefined",0)
@@ -550,16 +545,16 @@ myhud.add_txt("Defined",1)
 myhud.add_txt("Tool oriented",2)
 myhud.add_txt("",[1,2,3])
 myhud.add_txt("TWP-Orientation Vector X:",[1,2,3])
-myhud.add_pin("Xx: {:8.3f}","xyzacb-trsrn-gui.twp_xx",[1,2,3])
-myhud.add_pin("Xy: {:8.3f}","xyzacb-trsrn-gui.twp_xy",[1,2,3])
-myhud.add_pin("Xz: {:8.3f}","xyzacb-trsrn-gui.twp_xz",[1,2,3])
+myhud.add_pin("Xx: {:8.3f}","xyzbca-trsrn-gui.twp_xx",[1,2,3])
+myhud.add_pin("Xy: {:8.3f}","xyzbca-trsrn-gui.twp_xy",[1,2,3])
+myhud.add_pin("Xz: {:8.3f}","xyzbca-trsrn-gui.twp_xz",[1,2,3])
 myhud.add_txt("",[1,2,3])
 myhud.add_txt("TWP-Orientation Vector Z:",[1,2,3])
-myhud.add_pin("Zx: {:8.3f}","xyzacb-trsrn-gui.twp_zx",[1,2,3])
-myhud.add_pin("Zy: {:8.3f}","xyzacb-trsrn-gui.twp_zy",[1,2,3])
-myhud.add_pin("Zz: {:8.3f}","xyzacb-trsrn-gui.twp_zz",[1,2,3])
+myhud.add_pin("Zx: {:8.3f}","xyzbca-trsrn-gui.twp_zx",[1,2,3])
+myhud.add_pin("Zy: {:8.3f}","xyzbca-trsrn-gui.twp_zy",[1,2,3])
+myhud.add_pin("Zz: {:8.3f}","xyzbca-trsrn-gui.twp_zz",[1,2,3])
 myhud.add_txt("",[1,2,3])
-myhud.show_tags_in_pin("xyzacb-trsrn-gui.twp_status")
+myhud.show_tags_in_pin("xyzbca-trsrn-gui.twp_status")
 #/hud
 
 main(model, tooltip, work, size=4000, hud=myhud, lat=-60, lon=25)
