+++ configs/sim/axis/vismach/5axis/table-rotary_spindle-rotary-nutating/xyzacb-trsrn_twp/xyzacb-trsrn.ini
@@ -114,7 +114,7 @@ HALCMD = net twp-xy                      <= twp-helper-comp.twp-xy
 HALCMD = net twp-xz                      <= twp-helper-comp.twp-xz
 
 # load the optional vismach-gui and its hal connections
-HALCMD = loadusr -W ./xyzacb-trsrn-gui.py
+HALCMD = loadusr -W ../vismach/xyzacb-trsrn-gui.py
 # vismach machine signals
 HALCMD = net  :axis-x            joint.0.pos-fb             xyzacb-trsrn-gui.axis_x
 HALCMD = net  :axis-y            joint.1.pos-fb             xyzacb-trsrn-gui.axis_y
