+++ configs/sim/axis/vismach/5axis/table-rotary_spindle-rotary-nutating/xyzbca-trsrn_twp/xyzbca-trsrn.ini
@@ -111,7 +111,7 @@ HALCMD = net twp-xy                      <= twp-helper-comp.twp-xy
 HALCMD = net twp-xz                      <= twp-helper-comp.twp-xz
 
 # load the optional vismach-gui and its hal connections
-HALCMD = loadusr -W ./xyzbca-trsrn-gui.py
+HALCMD = loadusr -W ../vismach/xyzbca-trsrn-gui.py
 
 HALCMD = net  :axis-x            joint.0.pos-fb             xyzbca-trsrn-gui.axis_x
 HALCMD = net  :axis-y            joint.1.pos-fb             xyzbca-trsrn-gui.axis_y
